fix duplicate point test, and casts for fmod() - should fix windows build

+++ src/Lib/Geometry/util.cxx
@@ -503,7 +503,7 @@ makePolygonsTP (const Line &line, double width, poly_list& polys, texparams_list
             tp.set_minv(last_end_v);
             tps.push_back(tp);
 
-            last_end_v = 1.0f - (fmod( (dist - last_end_v), 1.0f ));
+            last_end_v = 1.0f - (fmod( (double)(dist - last_end_v), (double)1.0f ));
         }
 
         prev_outer = cur_outer;
